requirements_pythonanywhere.txt
Flask==3.0.0
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-Mail==0.9.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
Werkzeug==3.0.1
reportlab==4.0.7
Pillow==10.0.0